refactoring test structure
